+++ src/utils/types.ts
@@ -1,11 +1,9 @@
-import type { TRootState } from '../index'
+import type { TRootState } from '../services/store'
 import type { TAppActions } from '../services/actions'
-import { Dispatch, Action, ActionCreator } from 'redux'
+import { Action, ActionCreator } from 'redux'
 import { ThunkAction } from 'redux-thunk'
 
 export type TAppThunk<TReturn = void> = ActionCreator<ThunkAction<TReturn, Action, TRootState, TAppActions>>
-
-export type TAppDispatch = Dispatch<TAppActions>
 
 interface IIngredientType {
   readonly _id: string
